Inverter in voltage control mode 1 kW output
Inverter stage in a "good enough" state to build the DC-DC stage.
The inverter is operating in a simple closed-loop voltage control mode using PID control.
The interfaces are setup so that the gate driver is on the control board and is expecting a current reference. Note that this current reference does not work yet!